requirements.txt
appnope==0.1.3
asgiref==3.5.1
asttokens==2.0.8
async-generator==1.10
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bootstrap4==0.1.0
cachetools==5.2.0
certifi==2022.9.14
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.1
cryptography==38.0.1
decorator==5.1.1
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==3.2.12
django-allauth==0.50.0
django-appconf==1.0.5
django-bootstrap-v5==1.0.11
django-cors-headers==3.13.0
django-crontab==0.7.1
django-extensions==3.1.5
django-filter==22.1
django-imagekit==4.1.0
django-rest-framework==0.1.0
django-seed==0.3.1
django_debug_toolbar==3.8.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.0.4
executing==1.0.0
Faker==15.1.1
google-api-core==2.10.1
google-api-python-client==2.61.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.3
google-images-download==2.8.0
googleapis-common-protos==1.56.4
h11==0.14.0
httplib2==0.20.4
idna==3.4
ipython==8.5.0
jedi==0.18.1
matplotlib-inline==0.1.6
multidict==6.0.2
oauthlib==3.2.1
outcome==1.2.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pilkit==2.0
Pillow==9.1.0
prompt-toolkit==3.0.31
protobuf==4.21.6
psycopg2-binary==2.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
Pygments==2.13.0
PyJWT==2.4.0
PyMySQL==1.0.2
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.6.1
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.5.0
tomli==2.0.1
toposort==1.7
traitlets==5.4.0
trio==0.22.0
trio-websocket==0.9.2
uritemplate==4.1.1
urllib3==1.26.12
wcwidth==0.2.5
wsproto==1.2.0
yarl==1.8.1
gunicorn==20.1.0

